requirements.txt
## This file lists the dependencies for the project. To install them, run:
## pip install -r requirements.txt

pymupdf
reportlab
pandas
python-dotenv
sentence-transformers
faiss-cpu
numpy
requests
streamlit